requirements.txt
sphinx
pydata-sphinx-theme
myst-parser
sphinx-copybutton
sphinx-design
sphinx_rtd_theme
furo
sphinx_book_theme
sphinxcontrib-mermaid
graphviz      # 👈 llibreria Python per Graphviz
pillow